% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAseq.multi.DEG.R
\name{BMAseq.multi.DEG}
\alias{BMAseq.multi.DEG}
\title{BMAseq.multi.DEG}
\usage{
BMAseq.multi.DEG(
  dat.pheno,
  model.space,
  post.modelprob,
  var.pool,
  interact = FALSE,
  ind.incl.add = NULL,
  cut.FDR = 0.05
)
}
\arguments{
\item{dat.pheno}{Phenotypic data matrix with new interaction variables if interact=T.}

\item{model.space}{The model space including all possible models.}

\item{post.modelprob}{Posterior model probability for each gene.}

\item{var.pool}{Variables of interest, a vector.}

\item{interact}{Whether interaction terms are considered.}

\item{ind.incl.add}{Indices of additional included model.}

\item{cut.FDR}{False discovery rate criterion for identifying DE genes, default value is 0.05.}
}
\value{
A list consisting of
\item{post.incl.modelprob.Main}{Posterior inclusive model probability for each gene associated with main effect of each variables.}
\item{post.incl.modelprob.Interaction}{Posterior inclusive model probability for each gene associated with interaction effect of each variables, will be output if interact=TRUE.}
\item{post.incl.modelprob.MainInteraction}{Posterior inclusive model probability for each gene associated with main or interaction effect of each variables, will be output if interact=TRUE.}
\item{post.incl.modelprob.add}{Posterior inclusive model probability for each gene associated with additional inclusive models, will be output if ind.incl.add != Null.}
\item{index.incl}{Indices of the inclusive models.}
\item{eFDR.Main}{Estimated FDR for each gene associated with main effect of each variables.}
\item{eFDR.Interaction}{Estimated FDR for each gene associated with interaction effect of each variables, will be output if interact=TRUE.}
\item{eFDR.MainInteraction}{Estimated FDR for each gene associated with main or interaction effect of each variables, will be output if interact=TRUE.}
\item{eFDR.add}{Estimated FDR for each gene associated with additional inclusive models, will be output if ind.incl.add != Null.}
\item{indicator.eFDR.Main}{Indicators for DE genes associated with main effect of each variables.}
\item{indicator.eFDR.Interaction}{Indicators for DE genes associated with interaction effect of each variables, will be output if interact=TRUE.}
\item{indicator.eFDR.MainInteraction}{Indicators for DE genes associated with mean or interaction effect of each variables, will be output if interact=TRUE.}
\item{indicator.eFDR.add}{Indicators for DE genes associated with additional inclusive models, will be output if ind.incl.add != Null.}
\item{summary.nDEG}{A summary table of the number of identified DE gene associated with each variable.}
\item{summary.nDEG.add}{A summary table of the number of identified DE gene associated with additional inclusive models, will be output if ind.incl.add != Null.}
\item{DEG.bestmodel.Main}{DE genes associated with main effect of each variables and the best model used to identify each DE gene.}
\item{DEG.bestmodel.Interaction}{DE genes associated with interaction effect of each variables and the best model used to identify each DE gene, will be output if interact=TRUE.}
\item{DEG.bestmodel.MainInteraction}{DE genes associated with main or interaction effect of each variables and the best model used to identify each DE gene, will be output if interact=TRUE.}
\item{bestmodel.DEG.Main}{The best models used to identify DE genes associated with main effect of each variable and the DE gene identified by each model.}
\item{bestmodel.DEG.Interaction}{The best models used to identify DE genes associated with interaction effect of each variable and the DE gene identified by each model, will be output if interact=TRUE.}
\item{bestmodel.DEG.MainInteraction}{The best models used to identify DE genes associated with main and interaction effect of each variable and the DE gene identified by each model, will be output if interact=TRUE.}
}
\description{
DE genes identification with obtained posterior model probability for each gene in multivariate analysis using BMAseq approach
}
\details{
DE genes identification with obtained posterior model probability for each gene in multivariate analysis on RNA-seq count data using BMAseq approach
}
\author{
Lingsong Meng
}
