% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAseq.multi.postprob.R
\name{BMAseq.multi.postprob}
\alias{BMAseq.multi.postprob}
\title{BMAseq.multi.postprob}
\usage{
BMAseq.multi.postprob(
  dat.expr.counts,
  dat.pheno,
  var.pool,
  max.nvar,
  interaction = NULL,
  cut.BF = 1
)
}
\arguments{
\item{dat.expr.counts}{RNA-seq count data matrix (rows for genes and columns for subjects).}

\item{dat.pheno}{Phenotypic data matrix (rows for subjects and columns for variables).}

\item{var.pool}{Variables of interest, a vector.}

\item{max.nvar}{The maximum number of variables in a model.}

\item{interaction}{Variables considered in interaction terms, a vector, default is NULL.}

\item{cut.BF}{Bayes factor criterion, default value is 1.}
}
\value{
A list consisting of
\item{dat.expr.logcpm}{Normalized RNA-seq data matrix (rows for genes and columns for subjects).}
\item{weights}{Estimated voom weights.}
\item{dat.pheno.new}{Phenotypic data matrix including new interaction variables, will be same as input dat.pheno if interaction=NULL.}
\item{model.space}{Model space including all possible models.}
\item{post.modelprob}{Posterior model probability for each gene.}
}
\description{
Calculation of posterior model probability for each gene in multivariate analysis using BMAseq approach
}
\details{
Calculation of posterior model probability for each gene in multivariate analysis on RNA-seq count data using BMAseq approach
}
\author{
Lingsong Meng
}
