% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAseq.multi.R
\name{BMAseq.multi}
\alias{BMAseq.multi}
\title{BMAseq.multi}
\usage{
BMAseq.multi(
  dat.expr.counts,
  dat.pheno,
  var.pool,
  max.nvar,
  interaction = NULL,
  cut.BF = 1,
  cut.FDR = 0.05
)
}
\arguments{
\item{dat.expr.counts}{RNA-seq count data matrix (rows for genes and columns for subjects).}

\item{dat.pheno}{Phenotypic data matrix (rows for subjects and columns for variables).}

\item{var.pool}{Variables of interest, a vector.}

\item{max.nvar}{The maximum number of variables in a model.}

\item{interaction}{Specific interaction terms with input format 'A&B', default is Null.}

\item{cut.BF}{Bayes factor criterion, default value is 1.}

\item{cut.FDR}{False discovery rate criterion for identifying DE genes, default value is 0.05.}
}
\value{
A list consisting of
\item{dat.expr.logcpm}{Normalized RNA-seq data matrix (rows for genes and columns for subjects).}
\item{weights}{Estimated voom weights.}
\item{dat.pheno.new}{Phenotypic data matrix including new interaction variables, will be same as input dat.pheno if interaction is NULL.}
\item{model.space}{Model space including all possible models.}
\item{post.modelprob}{Posterior model probability for each gene.}
\item{post.incl.modelprob.Main}{Posterior inclusive model probability for each gene associated with main effect of each varibles.}
\item{post.incl.modelprob.Interaction}{Posterior inclusive model probability for each gene associated with interaction effect of each varibles, will be output if interaction is not NULL.}
\item{post.incl.modelprob.MainInteraction}{Posterior inclusive model probability for each gene associated with main or interaction effect of each varibles, will be output if interaction is not NULL.}
\item{summary.nDEG}{A summary table of the number of identified DE gene associated with each variable.}
\item{DEG.bestmodel.Main}{DE genes associated with main effect of each variables and the best model used to identify each DE gene.}
\item{DEG.bestmodel.Interaction}{DE genes associated with interaction effect of each variables and the best model used to identify each DE gene, will be output if interaction is not NULL.}
\item{DEG.bestmodel.MainInteraction}{DE genes associated with main or interaction effect of each variables and the best model used to identify each DE gene, will be output if interaction is not NULL.}
\item{bestmodel.DEG.Main}{The best models used to identify DE genes associated with main effect of each variable and the DE gene identified by each model.}
\item{bestmodel.DEG.Interaction}{The best models used to identify DE genes associated with interaction effect of each variable and the DE gene identified by each model, will be output if interaction is not NULL.}
\item{bestmodel.DEG.MainInteraction}{The best models used to identify DE genes associated with main and interaction effect of each variable and the DE gene identified by each model, will be output if interaction is not NULL.}
}
\description{
Multivariate analysis using BMAseq approach
}
\details{
Multivariate analysis on RNA-seq count data using BMAseq approach
}
\author{
Lingsong Meng
}
