% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAseq.uni.R
\name{BMAseq.uni}
\alias{BMAseq.uni}
\title{BMAseq.uni}
\usage{
BMAseq.uni(dat.expr.counts, dat.pheno, var.pool, cut.BF = 1, cut.FDR = 0.25)
}
\arguments{
\item{dat.expr.counts}{RNA-seq count data matrix (rows for genes and columns for subjects).}

\item{dat.pheno}{Phenotypic data matrix (rows for subjects and columns for variables).}

\item{var.pool}{Variables of interest, a vector.}

\item{cut.BF}{Bayes factor criterion, default value is 1.}

\item{cut.FDR}{False discovery rate criterion for identifying DE genes, default value is 0.25.}
}
\value{
A list consisting of
\item{dat.expr.logcpm}{Normalized RNA-seq data matrix (rows for genes and columns for subjects).}
\item{weights}{Estimated voom weights.}
\item{eFDR}{Estimated false discovery rate matrix (rows for genes and columns for variables of interest).}
\item{nDEG}{The number of DE genes associated with each variable of interest.}
\item{DEG}{DE genes associated with each variable of interest.}
}
\description{
Univariate analysis using BMAseq approach
}
\details{
Univariate analysis on RNA-seq count data using BMAseq approach
}
\author{
Lingsong Meng
}
