% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFit.single.R
\name{lmFit.single}
\alias{lmFit.single}
\title{BMAseq.multi}
\usage{
lmFit.single(
  dat.expr.counts,
  dat.pheno,
  var.pool,
  max.nvar,
  interaction = NULL
)
}
\arguments{
\item{dat.expr.counts}{RNA-seq count data matrix (rows for genes and columns for subjects).}

\item{dat.pheno}{Phenotypic data matrix (rows for subjects and columns for variables).}

\item{var.pool}{Variables of interest, a vector.}

\item{max.nvar}{The maximum number of variables in a model.}

\item{interaction}{Interactions.}
}
\value{
A list consisting of
\item{dat.expr.logcpm}{Normalized RNA-seq data matrix (rows for genes and columns for subjects).}
\item{weights}{Estimated voom weights.}
\item{summary.lmFit}{Output results of single linear models}
}
\description{
Analysis using single model approach
}
\details{
Analysis on RNA-seq count data using single model approach
}
\author{
Lingsong Meng
}
